Add roster class and set members in class variable
